Fix roles lost when the header is missing from requests
Use case:

* Odoo requests itself asking for a report; no SSO header present
* Direct access to odoo bypassing the proxy (for debugging only); no SSO
  header present

In these situations, the roles are lost each time such a request
hits Odoo. This is specially noticeable when a report is printed:
Odoo requests itself, as we have no header, the roles are removed from
the user, so an access error is triggered during the report generation,
but when the user browse on Odoo (through the SSO proxy), the roles are
instantly back.

Now, the role changes will be applied only if the header is
present (empty or not). It makes sense because we can consider that
if the header is absent, it comes from a trusted environment. The
reverse proxy enforces the header for the untrusted world.