Reset last_http_header_role when invalid
There could be various reason last_http_header_role mismatch with role
lines.
Either the role were added or removed manually on the user or the role
codes could have been changed.

In worst cases it can lead to a user not able to recover his login.
Because the presence of last_http_header_role it will bypass the
role assignation and will fail with an error 500 due to access rights
on the first thing it tries to read.

To fix it we reset last_http_header_role when:

A change on the relation user <-> group is done
A change on the relation user <-> role is done

The reset is bypassed when the change is done by the remote user auth