[FIX] password_security: Correctly look at previous password history
If a user has previous passwords stored and the `password_history` is set to 0 to disable the check, the line of code would evaluate to:

```
rec_id.password_history_ids[0:-1]
```

The above returns all items in the list except for the first last entry. One expects the above to evaluate to:

```
rec_id.password_history_ids[0:0]
```

The above returns an empty list which would then correctly not raise a warning.