[IMP] auth_saml nicer handling of logout and autoredirect
When autoredirect is enabled, but some technical accounts have a normal
password authentication setup, it can be difficult to access them as
logging out of a session will trigger the autoredirect feature.

We change this by overriding the logout redirection so that it sets the
disable_autoredirect parameter -> this way the user will get redirected
to the normal signin screen.