[FIX] vault: Error thrown when attempting to remove a right from a vault with more than 2 defined rigths.
Before the patch, an error would occur when clicking on OK after warning about key re-encryption. This is because the public_key field of user rights other than the one initiating the deletion is not available.

After the patch, the public_key field is computed, and therefore, when re-encrypting, the error no longer occurs.