[FIX] vault: Error when reencripting
To reproduce the error, the steps would be:
1. Access a Vault with a significant number of records, and some of them have null-valued keys.
2. Remove a permission from the vault and re-encrypt the keys.
3. When you save, an error will be thrown due to a required field being empty, and any keys that haven't been saved will be lost.

With the changes applied in this patch, the aforementioned error no longer occurs, and the keys that are in an inconsistent state no longer cause the others to become corrupted. In addition to fixing this, we have added the feature to disable the re-encryption assistant buttons to avoid the impression that nothing is happening.