[FIX] vault: Secret generator is not saving new value
When a field has assigned a value and we try to change it by
using the secrets generator, the new value is not saved. It is
happening because the click event on wizard buttons is propagated
and this trigger the bus event RELATIONAL_MODEL:NEED_LOCAL_CHANGES.

By doing this changes the value generated is assigned to the field
without loosing the focus. So the value is saved.

Steps to reproduce the problem:

1. Create a Vault
2. Create an Entry for the Vault
3. Create a Field for the Entry
4. Assign a value
5. Edit the value using the button to generate secrets
6. Save the new secret

You will see that the value salected on wizard is
not saved.

TT49554