[FIX] password_security: Fix reset password when user doesn't have password_write_date value.
It happens that when you install the module, all the users that previously existed before that won't have a value set for the new field password_write_date.
And when the validation of minimal hours has passed between the last password write date and now is done, it gets an error of unsupported operands between a bool and timedelta.
With this change, it won't be necessary to validate the last password date if it's the first time reseting password, and when the password is changed it will correctly save the value on password_write_date and will be working normally afterwards.